[mining] Change inBlock to be an unordered set
Summary:
Currently `inBlock` is a `std::set<txiter, CompareIteratorByHash>`, which depends on having
a txiter (`indexed_transaction_set::nth_index<0>::type::iterator`).  In future changes to
package handling, we want to avoid being dependent on mempool data structures of `mapTx`.

This commit is a step in that direction.  We convert `std::set<txiter, CompareIteratorByHash>`
to an `std::unordered_set<TxId, SaltedTxidHasher>`.  Which allows us to tell if a TxId has
already been added to the block template.

Test Plan:
  make check && ./test/functional/test_runner.py

Reviewers: O1 Bitcoin ABC!, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D2865